requirements.txt
docopt
python-pptx
